Add files via upload
It is a Hangman Game .The program selects a random word, and the player guesses one letter at a time to uncover the word. 
Player Solves a Riddle:
The player is given a riddle and needs to guess the correct word based on the hints provided.

Hints to Solve the Riddle:
The player is given up to three hints to help solve the riddle.

Restriction on Answers:
Only strings (words) are valid answers. Integers (numbers) are not allowed.

Red Highlight for Import Statements:
You mentioned that "import statements are shown in red to help the player navigate."